+++ lib/index.js
@@ -1,5 +1,5 @@
-const Weather = require('./javascript/weather.js').default
-const BackgroundImage = require('./javascript/background-image.js').default
+const Weather = require('./javascripts/classes/weather.js').default
+const BackgroundImage = require('./javascripts/classes/background-image.js').default
 
 // ------ display signup form -----------
 document.getElementById('signupLink').addEventListener('click', displaySignupForm)
