+++ lib/javascripts/classes/background-image.js
