+++ lib/javascripts/classes/weather.js
